requirements.txt
anthropic>=0.25.0
openai>=1.30.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
rich>=13.0.0
